Report the entire hash in the "committed" message.
This concerns the "committed" message posted to Rietveld. Prior to
34504a15bb3e, this used the committer's own core.abbrev. Now it's using
a hard-coded 7. Per recent discussions on chromium-dev, 7 is
insufficient to uniquely identify commits in large repositories such as
ours. If any abbreviation is to be done, a 12-digit minimum should be
used, but there's no reason not to use the entire hash.

Review URL: https://codereview.chromium.org/592293002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@292164 4ff67af0-8c30-449e-8e8b-ad334ec8d88c