Don't check bucket permissions if --no-auth.
Checking bucket permissions takes ~400ms. We don't
need to do this if --no-auth because we know we
won't get a 403 and the 404 check will be handled
later when we try to actually download the file.

Also, remove the check for a null bucket. This
can't happen since we will throw a parser error
in the main function before we get to this code.

R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/772203002

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@293250 4ff67af0-8c30-449e-8e8b-ad334ec8d88c