Make --output and --sha1_filename work well together.
--sha1_filename unnecessarily forces the output to have
the same filename and be in the same directory. The code
in main already correctly sets the file name to the
sha1_filename minus the .sha1, so the only change is
to actually use the --output path the same way the
rest of the code does.

R=iannucci@chromium.org

Review URL: https://codereview.chromium.org/752803003

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@293251 4ff67af0-8c30-449e-8e8b-ad334ec8d88c