my_activity.py: Fix to work with non-ascii characters
The default encoding for python 2.7 strings is ascii.  Some of
the titles of issues, such as http://crbug.com/477941, contain
non-ascii characters, and cause my_activity.py to throw an
exception.  Re-encode the unicode title string to ascii with
the 'replace' option, to replace non-ascii characters with '?'.

R=maruel@chromium.org
BUG=none
TEST=Run `my_activity.py` with #477941 in my issues list; correctly
     lists issues without throwing an exception.

Review URL: https://codereview.chromium.org/1151053003

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@295425 4ff67af0-8c30-449e-8e8b-ad334ec8d88c