Fix "git cl map-branches -vvv"
CL https://codereview.chromium.org/1172223002 changed the meaning of the
third value in the tuple returned by get_cl_statuses from a color to the
status of the branch and broke the formatting of output.

Convert the status to the color using color_for_status helper function.

BUG=499189

Review URL: https://codereview.chromium.org/1175103004

git-svn-id: http://src.chromium.org/chrome/trunk/tools/depot_tools@295618 4ff67af0-8c30-449e-8e8b-ad334ec8d88c